refactor: move export and importPreview to dynamic query
